make the lore dump exeuction node wrapper generic
