+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/lore/dump.scala
@@ -16,16 +16,19 @@
 
 package com.nvidia.spark.rapids.lore
 
-import com.nvidia.spark.rapids.{DumpUtils, GpuColumnVector, KudoSerializedTableColumn, SerializedTableColumn}
+import com.nvidia.spark.rapids.{DumpUtils, GpuColumnVector, GpuExec, KudoSerializedTableColumn, SerializedTableColumn}
 import com.nvidia.spark.rapids.GpuCoalesceExec.EmptyPartition
 import com.nvidia.spark.rapids.jni.kudo.KudoSerializer
 import com.nvidia.spark.rapids.lore.GpuLore.pathOfChild
+import com.nvidia.spark.rapids.shims.ShimUnaryExecNode
 import org.apache.hadoop.fs.Path
 
 import org.apache.spark.{Partition, SparkContext, TaskContext}
 import org.apache.spark.broadcast.Broadcast
 import org.apache.spark.rdd.RDD
+import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Attribute
+import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.rapids.execution.GpuBroadcastHelper
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.vectorized.ColumnarBatch
@@ -115,4 +118,33 @@ class SimpleRDD(_sc: SparkContext, data: Broadcast[Any], schema: StructType) ext
   }
 
   override protected def getPartitions: Array[Partition] = Array(EmptyPartition(0))
+}
+
+/**
+ * A reusable wrapper execution node that adds LoRE dumping capability to any SparkPlan
+ * that doesn't follow GpuExec's internalDoExecuteColumnar protocol.
+ *
+ * This can be used by any execution context that needs to wrap a child plan with LoRE
+ * dumping functionality, not just data writing commands.
+ */
+case class GpuLoreDumpWrapper(
+    child: SparkPlan,
+    loreDumpInfo: LoreDumpRDDInfo)
+    extends ShimUnaryExecNode with GpuExec {
+
+  override def output: Seq[Attribute] = child.output
+
+  override protected def doExecute(): RDD[InternalRow] = {
+    throw new UnsupportedOperationException(
+      s"${getClass.getCanonicalName} does not support row-based execution")
+  }
+
+  override protected def internalDoExecuteColumnar(): RDD[ColumnarBatch] = {
+    val childRDD = child.executeColumnar()
+    val rdd = new GpuLoreDumpRDD(loreDumpInfo, childRDD)
+    rdd.saveMeta()
+    rdd
+  }
+
+  override def nodeName: String = s"LoreDump(${child.nodeName})"
 }
